pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "py-cqrs-ddd-auth"
version = "0.1.0"
description = "A toolkit-native authentication and authorization library built using CQRS, DDD, and Saga patterns"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Myron Gourlis" }
]
keywords = ["authentication", "authorization", "cqrs", "ddd", "keycloak", "oauth2", "jwt", "2fa", "abac"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Security",
    "Typing :: Typed",
]

# Core dependencies
dependencies = [
    "py-cqrs-ddd-toolkit @ git+ssh://git@github.com/mgourlis/py-cqrs-ddd-toolkit.git",
    "search_query_dsl @ git+ssh://git@github.com/mgourlis/search_query_dsl.git",
    "stateful-abac-common @ git+ssh://git@github.com/mgourlis/stateful-abac-policy-engine.git#subdirectory=common",
    "stateful-abac-sdk[db] @ git+ssh://git@github.com/mgourlis/stateful-abac-policy-engine.git#subdirectory=python-sdk",
    "pyotp>=2.9.0",  # TOTP/HOTP for 2FA
]

[project.optional-dependencies]
di = [
    "dependency-injector>=4.41.0",
]
keycloak = [
    "python-keycloak>=3.0.0",
    "python-jose[cryptography]>=3.3.0",
    "httpx>=0.25",
]
fastapi = [
    "fastapi>=0.100",
    "pydantic>=2.0",
]
django = [
    "django>=4.2",
]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-cov>=4.1",
    "httpx>=0.25",
    "ruff>=0.1.0",
    "pre-commit>=3.5.0",
]
all = [
    "py-cqrs-ddd-auth[keycloak,fastapi,django,dev,di]",
]

[project.urls]
Homepage = "https://github.com/mgourlis/py-cqrs-ddd-auth"
Documentation = "https://github.com/mgourlis/py-cqrs-ddd-auth#readme"
Repository = "https://github.com/mgourlis/py-cqrs-ddd-auth"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/cqrs_ddd_auth"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
pythonpath = ["src"]
addopts = "--cov=src/cqrs_ddd_auth --cov-report=term-missing"

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
